login + sign in

+++ demo/src/main/java/prototype/example/demo/repository/FormulaRepository.java
@@ -3,8 +3,10 @@ package prototype.example.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import prototype.example.demo.entity.Formula;
 
+
+import java.util.List;
 import java.util.Optional;
 
-public interface FormulaRepository extends JpaRepository<Formula, Integer> {
+public interface FormulaRepository extends JpaRepository<Formula, Long> {
     Optional<Formula> findByName(String name);
-} 
+}
